Rewrite UI in YAD instead of Zenity

+++ plugins/repo/remove_repo.sh
@@ -1,15 +1,20 @@
-# Name: Remove repositories
-# Command: list_repos
+# Name: Delete repositories
+# Command: delete_repos
 
-list_repos() {
-build_repolist "/etc/yum.repos.d/" "false"
-while repos=$(zenity --list --checklist --width=900 --height=600 --title="Remove repositories" --text="Removing a repo will delete the repo from the system. It does not affect installed packages." --hide-header --column "Select" --column "Name" --column "Description" --column "Status" --ok-label="Remove selected" --cancel-label="Back" "${repolist[@]}"); do
-	selrepo=$(echo $repos | tr "|" "\n")
-	for repofile in $selrepo; do
-		show_msg "Removing $repofile"
-		rm -f /etc/yum.repos.d/$repofile
-		[[ ! -f /etc/yum.repos.d/$repofile ]]; exit_state
-	done
-	build_repolist "/etc/yum.repos.d/" "false"
+delete_repos() {
+while :
+do
+    build_repolist "/etc/yum.repos.d/" "false"
+    repos=$(dialog_show --list --checklist --width=900 --height=600 --title="Select repositories to delete" --text="Deleting a repo will delete the repo from the system. It does not affect installed packages." --no-headers --print-column="2" --column "Select:CHK" --column "Name" --column "Description" --column "Status" --button="Back:1" --button="Delete selected:0" "${repolist[@]}")
+    if [[ $? -eq 0 ]]; then
+	    selrepo=$(echo $repos | tr "|" "\n")
+	    for repofile in $selrepo; do
+		    show_msg "Deleting $repofile"
+		    rm -f /etc/yum.repos.d/$repofile
+		    [[ ! -f /etc/yum.repos.d/$repofile ]]; exit_state
+	    done
+	else
+	    break
+	fi
 done
 }
